+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_returning_widgets/config_parser.dart
@@ -4,18 +4,12 @@ class _ConfigParser {
   static const _ignoredNamesConfig = 'ignored-names';
   static const _ignoredAnnotationsConfig = 'ignored-annotations';
 
-  static const _defaultIgnoredAnnotations = [
-    'FunctionalWidget',
-    'swidget',
-    'hwidget',
-  ];
-
   static Iterable<String> getIgnoredNames(Map<String, Object> config) =>
       _getIterable(config, _ignoredNamesConfig) ?? [];
 
   static Iterable<String> getIgnoredAnnotations(Map<String, Object> config) =>
       _getIterable(config, _ignoredAnnotationsConfig) ??
-      _defaultIgnoredAnnotations;
+      functionalWidgetAnnotations;
 
   static Iterable<String>? _getIterable(
     Map<String, Object> config,