+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_static_class/examples/correct_ignore_names_example.dart
@@ -1,0 +1,3 @@
+final someRiverpodProvider = 42;
+
+final useSomeRiverpodHook = 42;