+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_static_class/examples/correct_ignore_private_example.dart
@@ -1,0 +1,7 @@
+void _privateGlobalFunction() {}
+
+var _privateGlobalVariable = 42;
+
+final _privateGlobalFinalVariable = 42;
+
+const _privateGlobalConstant = 42;